+++ src/Components/common/Modal/ImageModal.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled, { css } from 'styled-components';
-import useIsWideView from 'Components/PCNav/useIsWideView';
+import useIsWideView from 'Components/SideNav/useIsWideView';
 
 const ImageModal = ({ setIsModalOn, imgSrc }) => {
   const isWideView = useIsWideView();
