+++ src/Pages/Chat/Chat.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ChatList from 'Components/Chat/ChatList';
 import PCChat from 'Components/Chat/PCChat';
-import useIsWideView from 'Components/PCNav/useIsWideView';
+import useIsWideView from 'Components/SideNav/useIsWideView';
 
 const Chat = () => {
   const isWideView = useIsWideView();
